Organize public utilities into a utils/ package (Slice 0)
Replace the pre-2.0 scatter (ttkbootstrap.utility grab-bag +
ttkbootstrap.colorutils) with a first-class ttkbootstrap.utils package:

  utils/color.py     <- colorutils.py (color-model conversion/manipulation)
  utils/scaling.py   <- enable_high_dpi_awareness, scale_size (from utility.py)
  utils/platform.py  <- windowing_system (from utility.py)
  utils/__init__.py  re-exports the whole surface from one namespace

The old module paths become thin warn-and-forward shims (same pattern as
ttkbootstrap.publisher): importing ttkbootstrap.utility / ttkbootstrap.colorutils
emits a one-time DeprecationWarning and re-exports the moved names (incl. the
colorutils model constants and the two already-internal utility forwards).
Removed in 3.0. The top-level re-exports (ttk.scale_size, ttk.contrast_color,
...) are unchanged.

All first-party callers are migrated onto ttkbootstrap.utils so importing
ttkbootstrap stays warning-free (window, style/theme, meter, tooltip, tableview,
toast, floodgauge, scrolled, and the color/font/date/base dialogs). The private
ttkbootstrap.internal.utility (get_image_name/center_on_parent) is untouched.

This is the home the later slices (typography, deferred-config) are born into.
No behavior changed -- the functions are the same objects.

Tests: rewrite test_utility_api onto ttkbootstrap.utils (+ utils-package surface
+ shim-warns-but-forwards coverage, made order-independent via a forced
re-import); update test_scaling's import and the window.py source-order strings
(utility. -> utils.). New breaking-changes entry (Deprecated old paths, no break).
