+++ tests/test_utility_api.py
@@ -1,11 +1,18 @@
 """Headless tests for the 2.0 public-utility surface.
 
-Covers that the utility/colorutils helpers are reachable at the top level
-(`ttk.<name>`) and that the new `windowing_system` helper matches the raw Tcl
-call it consolidates.
+Covers that the color / scaling / platform helpers are reachable at the top
+level (`ttk.<name>`) and through the canonical `ttkbootstrap.utils` package,
+that `windowing_system` matches the raw Tcl call it consolidates, and that the
+old `ttkbootstrap.utility` / `ttkbootstrap.colorutils` module paths still work
+as warn-and-forward shims (Slice 0).
 """
+import importlib
+import sys
+
+import pytest
+
 import ttkbootstrap as ttk
-from ttkbootstrap import colorutils, utility
+from ttkbootstrap import utils
 
 
 # --------------------------------------------------------------------------
@@ -25,11 +32,40 @@ def test_public_utilities_are_top_level_and_in_all():
         assert name in ttk.__all__, f"{name} missing from ttk.__all__"
 
 
-def test_top_level_names_are_the_module_objects():
+def test_utils_package_exposes_the_full_surface():
+    for name in _UTILITY_NAMES + _COLOR_NAMES:
+        assert hasattr(utils, name), f"utils.{name} is missing"
+        assert name in utils.__all__, f"{name} missing from utils.__all__"
+
+
+def test_top_level_names_are_the_utils_objects():
     # the re-exports must be the real functions, not shadowing copies
-    assert ttk.scale_size is utility.scale_size
-    assert ttk.windowing_system is utility.windowing_system
-    assert ttk.contrast_color is colorutils.contrast_color
+    assert ttk.scale_size is utils.scale_size
+    assert ttk.windowing_system is utils.windowing_system
+    assert ttk.contrast_color is utils.contrast_color
+
+
+# --------------------------------------------------------------------------
+# back-compat shims (ttkbootstrap.utility / ttkbootstrap.colorutils)
+# --------------------------------------------------------------------------
+
+def test_legacy_utility_module_warns_but_forwards():
+    # drop any cached module so the module-level warning re-fires regardless of
+    # test order
+    sys.modules.pop("ttkbootstrap.utility", None)
+    with pytest.warns(DeprecationWarning, match="moved to ttkbootstrap.utils"):
+        legacy_utility = importlib.import_module("ttkbootstrap.utility")
+    assert legacy_utility.scale_size is utils.scale_size
+    assert legacy_utility.windowing_system is utils.windowing_system
+
+
+def test_legacy_colorutils_module_warns_but_forwards():
+    sys.modules.pop("ttkbootstrap.colorutils", None)
+    with pytest.warns(DeprecationWarning, match="moved to ttkbootstrap.utils"):
+        legacy_colorutils = importlib.import_module("ttkbootstrap.colorutils")
+    assert legacy_colorutils.contrast_color is utils.contrast_color
+    # the model constants (not in __all__) still resolve through the shim
+    assert legacy_colorutils.HEX == "hex" and legacy_colorutils.RGB == "rgb"
 
 
 # --------------------------------------------------------------------------
@@ -45,7 +81,7 @@ def test_windowing_system_returns_known_value(root):
 
 
 # --------------------------------------------------------------------------
-# scale_size + colorutils behave through the top-level names
+# scale_size + color helpers behave through the top-level names
 # --------------------------------------------------------------------------
 
 def test_scale_size_scales_a_pair(root):
@@ -53,6 +89,7 @@ def test_scale_size_scales_a_pair(root):
     assert isinstance(scaled, list) and len(scaled) == 2
 
 
-def test_colorutils_round_trip():
-    assert ttk.color_to_rgb("#ff5733", model=colorutils.HEX) == (255, 87, 51)
-    assert ttk.color_to_hex((255, 87, 51), model=colorutils.RGB) == "#ff5733"
+def test_color_helpers_round_trip():
+    from ttkbootstrap.utils.color import HEX, RGB
+    assert ttk.color_to_rgb("#ff5733", model=HEX) == (255, 87, 51)
+    assert ttk.color_to_hex((255, 87, 51), model=RGB) == "#ff5733"
